[[ ScriptifyLibURL ]] In order to make it easier to submit modifications to liburl, it has been scriptified and moved into the engine ide-support folder.
In order to achieve this with the minimum amount of potential issues the
revliburl button in revLibrary remains. In particular, this means that
the standalone builder will continue to operate unchanged.

The home stack's InitializeLibraries step has been modified so that:
  - it inserts the script of stack 'revliburl' into back.
  - it sets the script of button 'revliburl' of stack 'revlibrary' to
    that of the script of the revliburl stack.
  - it skips inserting the button revliburl script into back.

The home stack's stackfile list (on card 2) has been updated with a
reference to the revliburl script-only-stack.

The script of group 1 of revlibrary has been updated so that the
revLoadLibraries call is *not* made in development mode - this is
because it has already occurred in the home stack. i.e. This step
should only occur in a standalone.

The installer package.txt file has been updated to include the
reference to ide-support/revliburl.livecodescript.