[[ CmdLineBuilder ]] Skip linux-armv6-hf; make linux-x86_64 conditional
We're never going to make x86-64 Linux builds of 6.7.x available,
so only try and include x86-64 Linux binaries in non-6.7
installers.

Similarly, we're not currently planning to do regular builds of
the engine for armv6-hf (Raspberry Pi).  The armv6-hf builds are
skipped for now; future armv6-hf builds will need to comment back
in the relevant lines of package.txt or find a better solution.