[[ bug 9159 ]] Print keyboard shortcut doesn't work in Script Editor.
The default Script Editor preferences contained a small typo which linked the Print shortcut to command p', this meant that command p never triggered the Print Dialog from the Script Editor. The defaults have been updated and a check put in to check for the incorrect preference in the user's preferences and correct it if it is found.