[[ Frame Behavior ]] Clean up default preference options
With the generateFrame calls cleaned up, we can now add the default preference options
in the correct place (i.e. in generateFrame). This patch also extends the preference
controlling whether nav tabs display icons or labels to all palettes that show the
preference action (and removes it from the inspector behavior, where it is no longer
necessary)