[[ Property Data ]] Rework IDE object properties
Previously the data files for object types had their elements in a
different order to the main property definition file, which was less
than ideal.

This moves them around to have the same structure, and modifies the
IDE accordingly, unifying some of the property/object property code.

Also, all broken tests are fixed.