Don't delete stack files before overwriting
On Linux, the IDE would delete the target stack file, if it already
existed, before attempting to overwrite it using `save`.  This wasn't
correct, because `save` should implement the necessary truncation
itself, and because `save` might legitimately issue an error and fail
before modifying the target file (in which case it should be
preserved).