Support for test stack reuse
At the moment the extension builder creates a new stack each time the
user tests a widget etc which can be annoying if you want to script some
handlers or just add a few buttons to set properties. This commit adds
support for a stack named test.livecode to be placed in the extension
directory which if found will be loaded instead of creating a new stack.