[Perf][Bug 15588] Only update script editor styles if preferences don't match.
Users don't change their script editor display preferences very often,
and it's very slow to reset all the fonts of the script editor due to
need to completely re-layout the field.  For large scripts, only
resetting the editor field's styles when they don't match the current
preferences can save a significant amount of time.  This patch speeds
up switching tabs in the script editor by approximately 30%.

For example, on my system, switching to the revidelibrary script was
about 1.9 seconds and this patch reduces the time to about 1.1 s.