[Perf][Bug 15588] Align script editor handler list text more efficiently
Looping over the lines of the field and adjusting the text alignment
is quite slow -- instead, build the alignment information into the
generated htmlText directly.  This saves about 2/3 of the time needed
to update the left bar's handler list (0.4 seconds out of 0.6 seconds
when opening the IDE library script on my machine).