[Perf][Bug 15588] Cache script editor fields for object scripts
For very large scripts, resetting and reformatting the script field is
very costly -- sometimes over a second.  This is work that's largely
duplicative, especially when switching frequently between editor tabs
(which can occur because you have a script library open and you are
referring to its source code while writing a handler that makes use of
it).

By keeping editor fields for recent tabs around, invisibly, and
swapping them in and out as the user switches tabs, a substantial
speed up can be obtained when working with large scripts.

This patch hacks something into the script editor to achieve this.
Fortunately, it's relatively self-contained.  It currently caches a
maximum of 10 editor fields per script editor window, and makes the
field update cost of switching tabs negligible when there are no
script or layout changes required.