dictionary: Make crossrefs in description blocks look like normal text
Some dictionary entries have a _lot_ of cross references in their
description blocks, which can be extremely visually noisy.  Since all
the cross references in the description are already listed in the
dictionary entry's references section, it's okay to make the links in
the description block look like normal text until mouseover - it
doesn't hinder discoverability but it does make it easier to read.