[[ Tutorial ]] Make proper distinction between different types of stack reference This patch makes various highlight and wait actions work for different palette references in the tutorial system.
The previous special casing for the script editor's Apply button has been
generalised to any named component of any palette - it is then up to the
palettes to implement 'absoluteRectOfObject' for the relevant component name.

The 'to side' parameter has been removed from the tutorial stack positioning code
as the stack was only ever positioned below or above the target stack when the
target was the menubar. This logic has therefore been moved inside the position handler.

The name of the property inspector palette has been changed to 'inspector' to avoid
conflict with the 'highlight property' syntax.