Rename stack "DataViewControl"
The "DataViewControl" stack is used for the Project Browser. It is something I originally wrote and the name conflicts with my own stack. This commit changes the references to the stack from "DataViewControl" to "revDataViewControl". In addition to this commit, the following code must be executed in order to change stack files within the IDE.

Change DataView stack name
----------------------------------------
set the name of stack "DataViewControl" of stack "revCore" to "revDataViewControl"
save stack "revCore"

Change behaviors of Project Browser DataView
---------------------------------------------------------------
set the behavior of group "objectList" of card "list" of stack "revIDEProjectBrowser" to the long id of button "DataView Behavior" of stack "revDataViewControl"
set the behavior of button "dvTrackDragDrop" of card "list" of stack "revIDEProjectBrowser" to the long id of button "dvTrackDragDrop Behavior" of stack "revDataViewControl"
set the behavior of scrollbar "dvVScrollbar" of card "list" of stack "revIDEProjectBrowser" to the long id of button "VScrollbar Behavior" of stack "revDataViewControl"
set the behavior of group "dvDropIndicator" of card "list" of stack "revIDEProjectBrowser" to empty
save stack "revIDEProjectBrowser"