[Bug 16992] Make sure handling "saveStackRequest" cancels IDE save
According to the dictionary, handling the `saveStackRequest` message
(and failing to pass it) should cancel the `save` command.

The IDE calls the `save` command with messages locked, which prevents
the current stack from receiving the normal `saveStackRequest`
message.  The IDE therefore synthesizes one.

Previously, the IDE ignored whether or not the `saveStackRequest`
message was handled, and continued with the save operation regardless.
This prevented some useful applications of of intercepting the `save`
command (e.g. forcing the on-disk format to a specific stack format
version).  This patch ensures that the IDE save operation is cancelled
if the `saveStackRequest` message is blocked.