Correctly display password protected stacks in the Message Box
Display "pwd" instead of the number of script lines for password
protected stacks used as front scripts, back scripts or library
stacks when displayed in lists in the Message Box.

Add the revIDEObjectScriptLines function to the revIDELibrary. This
function returns the number of lines in the script of an object or
"pwd" if the object is on a password protected stack.