Show correct style and menuMode options for buttons in the PI
As menu buttons require both their style and menuMode properties to
be set the option to set the style of a button (push, default,radio
etc) to "menu" in the Property Inspector has been removed. Similarly
the style editor for menu buttons has been hidden and the menuMode
editor shown. Tab panels cannot have their menuMode or style set in
the Property Inspector.