Evaluate boolean expressions in the Message Box
The Message Box will now attempt to execute script as a boolean
expression, for example

there is a stack "test"

If the expression can be evaulated either globally or in the context
of the current stack the result is displayed.

Additionally the stack popup in the Message Box was not popping up
due to the layering of the controls in the Message Box. The properties
of the "Open Stacks" button have been changed to improve the
appearance and the layer of the "background" group is set to always
be at the bottom.