Update the Project Browser when a control is pasted into a group
The Project Browser receives the ideNewControl message when a control
is pasted into a group but the object ID passed with the message
does not include the fact that the new control is grouped. By using
the long ID of the passed object ID the Project Browser correctly updates
with the new, grouped control.