[[ Bug 17279 ]] Relativize appropriate stackFile paths
This patch adds a custom setprop handler for the stackFiles property
of stacks.

The custom setprop handler will relativize any file paths in the new
value relative to the target object in the case where the referenced
file is in the same folder, or in a folder beneath the folder containing
the target object.