Fix Message Box errors when setting background layer
In some cases the layer of the "background" group is set when
the stack is not open, this causes an error and is resolved by
attempting to set the layer within a try.

There was also an error the first time the Front Scripts tab was
opened, due to referring to a control in the footer than did not exist,
moving the reference to openCard fixed the error.