Prevent the IDE handling newBackground and deleteBackground messages
When a group is created newGroup and newBackground messages are sent,
the newBackground messgage is for compatibility with HyperCard. Both
messages are mapped to an ideNewControl message which was causing the
Project Browser to believe 2 new groups had been deleted and update
accordingly. Removing the newBackground message from the Message
Handler Library ensures only the newGroup message is handler.

The deleteBackground message has also been removed for the same reason.