Improve the appearance of the Project Browser
The margins of the field in each row template for the Project Browser
are set to standard margins and each row behavior has been updated to
lay out the controls taking the text size preference into account.

The revPreferencesGUI stack has been updated as it was resetting the
Project Browser Text Size preference to default when opened.