Show the correct state of the Select Grouped icon in the Menubar
On opening the Menubar sets the state for the Select Grouped icon
based on the value of the selectGroupedControls property, Previously
the icon was set based on not the selectGroupedControls, causing the
icon to appear in the wrong state.

The default icon for Select Grouped is off and the highlighted icon is
on, the images for these states were inverted, the image files have
been renamed so the correct icon is used for each state.