[Bug 16848] Force documentation HTML regeneration if missing
Sometimes the IDE documentation builder can get wedged -- it can think
that the documentation doesn't need regenerating, even though the
`api.html` file is missing from the documentation cache.

Sometimes, the documentation builder remains wedged even when the
whole `~/.runrev` directory is removed!

This patch changes the logic for ensuring the dictionary is available
to explicitly check for the presence (or otherwise) of the `api.html`
file.  Hopefully this will help unwedge some users dictionaries.