[[ ExternalDependencies ]] Support LCS based dependencies for externals
This patch provides support for externals found
in the Tools/Ext directory to include supporting
LCS based stackfile dependencies and for those
dependencies to manage dependencies between
themselves and other standalone inclusions.

In the course of development some unnecessary code
was removed from revDeployLibrary and revSBLibrary
was elevated to an always loaded IDE back script.
Additionally the recursive stackfile lister was
refactored.

There are dependencies between this patch and an
associated patch to the livecode repo.