Don't set IDE scripts that don't compile
Previously there was a check to ensure script editor
scripts were not set if they didn't compile. This has
been extended to all IDE scripts.

An appropriate warning is displayed if the script
editor or tab will be closed offering the chance to
save uncompilable work in progress if necessary.