Improve saving of stacks from script editor
This patch ensures that the user receives a
warning when trying to save an IDE script
from the script editor and one of the open
scripts from that mainstack has compilation
errors.

It also replaces revTargetStack with
ideMainStackOfObject to ensure all
open scripts within a stackFile are
applied during a save. Previously
editing the script of a mainstack
object and the script of a substack
object would not apply them both
before save.