[[ Bug 2867 ]] Workaround IDE hang when tracing resizeStack & moveStack
Setting the mode of the script editor to debug
must be sent in time to give the script editor
time to update. This unfortunately allows time
for the next resizeStack or moveStack message
to be handled resulting in an IDE hang and eventual
crash. This has been an issue for a long time.

This patch ignores errors in those two handlers
as a better alternative to an IDE hang. Ideally
the script editor could be reworked in the future
to not require the send in time.