[[ Bug ]] ensure iconGravity is empty when creating buttons in IDE
Previously the iconGravity was not being copied
during button creation which had the effect of
iconGravity always being empty on new buttons.

When generating the UI for revTools and revMenubar
the iconGravity of the templateButton was being
set to top but as the iconGravity from the
templateButton wasn't being copied by the engine
the UIs were designed with iconGravity empty.

This patch also ensures the templateButton is reset
before generating revMenubar