[Bug 18290] Use datagrid template safely while building headers
This patches fixes an unitialised variable use during datagrid
custom header initialisation.  The fix ensures that the ID of the
template stack is retrieved safely, and that custom column header
construction only proceeds if both a template stack ID is available
and a custom template is enabled for the column.