[Debugger] Implement remote debugger
This patch implements a remote debugger mode in the debugger
library. The library is used both in the standalone and the
IDE and provides a common API to both.

The debugger library depends on the common library and the
metadata library so they both need to be included in the
standalone.

This patch also moves watch and breakpoint init when a
stack opens into the debugger library.