[[ Cleanup ]] revtemplatechooser is no longer used
The revtemplatechooser stack appears to be an artifact from the revMedia days. It's not referenced from anywhere and the buttons all have code that explicitly does nothing. While it would be nice someday to have this sort of functionality as templates for new users (especially pluggable/maintainable), today the samples seem to serve the same purpose.