[[ Bug 19480 ]] Overhaul message box execution This patch simplifies message box execution by relying exclusively on compile checking to find the correct form of the script to execute, and always using the `send <script blob> to obj` form to execute (in non-debug mode).
The message box has been changed to always display the (single)
context object, either the active stack, selectedObject: <obj>, or
mouseControl <obj>. Thus it is always explicit what the context of
message box execution is.

This also fixes bug 19480 since the latter was caused by the script
being executed a second time after throwing an execution error in
the context of the intelligence object.