[Bug 19229] Improve project browser object ref robustness
The project browser uses long id as an object reference. This results
in a number of issues because when a stack is saved for the first
time the long id of its objects changes to include the filename.

This patch checks the long id and a ruggedised form of the long id
when getting the object row. A much more invasive patch should be
considered for a major release where an opaque object reference
be made available that is robust to name, filename and id changes
and that used instead of long id.