[[ Bug 20258 ]] Improve control structure completion
This patch adds control structure commpletion for `switch`
`case` to be completed with `break`. This change will reduce
continuation errors in user scripts.

This patch also adds control structure completion for `if` when
the last token of the line is `then`. Due to the numerous variants
of `if` this was the only variant with a low risk of regressions.