[[Bug 20289]] Improve dictionary sort
Initial complaint was that "me" is very difficult to locate in the dictionary.  The current priority search ends up with terms in an unsorted state since it does not explicitly maintain the alphabetical sort.  Additionally, the prioritized sort is looking at all syntax terms which is probably more than is needed for the priority portion.

Added a new regex for an exact word match.  This will be the first one used for the sort and will use the display name.  This will allow for things like "me" to appear at the top and other short words to be closer to the top.  The second regex will use the existing "start of word" search, but also only search the display name.  Finally, the display names will be compared to maintain alphabetical order.