[[ Load Components ]] Load stacks from commercial components
This patch allows the IDE to load stacks from component folders in
the commercial repository. It also vastly simplifies the logic in
the initialisation library by depending on libraries to be included
in the stackFiles of stack home meaning there is only a single point
where stackFile paths need to be determined. Because of the change
in logic the ide library needed to be special cased to allow loading
the old ide library which still exists on this branch.