[[ Tests ]] Test script compilation
This patch tests the scripts in the IDE compile and checks for
explicit variable compliance. In order to avoid masses of unhelpful
output the test does not assert explicit variable compliance if the
script does not compile.

This patch removes a code sample of a `ceiling` function from the
resources stack as this no longer compiles given the function of the
same name in the engine.