[[ Bug 18739 ]] Dictionary auto-search on first char freezing cursor
Did some preliminary timing at various points of the js.  The actual search `tData = dataSearch(pTerm);` was not the primary driver of the amount of time taken between typing a character and presentation of the information.  The search was in the 100ms range but the HTML generation `$("#table_list").html(tHTML);` would get up to 500ms frequently.  After researching methods to improve performance, the idea of delaying the actual search stood out as a preferred method for this use case.

Implemented a function (debounce) so that the search is not performed until 250ms after the last character is typed.  This would add 250ms to a single character search, but will more than overcome that when multiple characters are typed in quick succession.  If looking for something like "#", then the extra time is barely noticeable.  If typing multiple letters (especially 3 or more), the performance improvement is very evident.

The code is based on the following post:
https://stackoverflow.com/questions/12538344/asynchronous-keyup-events-how-to-short-circuit-sequential-keyup-events-for-speed
The actual lines of code are cut down from a function included in this project:
https://github.com/jashkenas/underscore
The project's code is under a MIT License