[[ Bug 20289 ]] Add term match to sort function
Per an earlier suggestion, added another regex to the sort function.  Before the alpha sort, added the regex to put term matches within the display name ahead of entries that just had term matches within the syntax.  When checking the letter "t", it increased the search time from around 50ms to 68ms.

Updated sort function to use a single pair of variables for the regex tests and declared them at the top of the function.  This was done to make it easier to add/subtract regex tests in the future.

Updated the spacing in the code to be consistent with the rest of the file.  In this file, tab characters are used vice the normal 4 spaces for indentation that are specified in the style guide.

Found a console.log entry that had not been commented out and corrected it.