[[ Live Errors ]] Improve live errors for explicit vars
This patch improves live errors when the user has `the explictVariables`
preference `true`. Previously the compilation done to generate the script
description was done with `explicitVariables` set the `false`. This patch
changes that to attempt to compile with the user's preferred setting and
if that fails and the `explicitVariables` was `true` to try again without
it in order to extract the script description while retaining the errors
from the user setting. This won't be necessary if the compiler is changed
to allow a script description to be generated even if the compilation had
errors.