[[ Bug 20145 ]] Fix issues with breakpoint movement on edit
This patch fixes a couple of issues with breakpoint movement and removal
when editing scripts. Additionally it uses more efficient code to determine
if the user was editing the line of a breakpoint with only whitespace
before the offset.

With this patch breakpoints should:

- move down correctly when the insertion point is on the beginning of the line
(such as `end Handler`)
- move up correctly if te user deletes part of (or before) the line they are on
and when moved up it is still valid to place a breakpoint on the line