[[ Bug 20392 ]] Eat closing brackets when typing
Because we have bracket completion and people are used to typing
closing brackets it saves time if when typing a closing bracket
that we look ahead and if the next char is already that closing
bracket then we just replace it rather than inserting another.
This means the selection will now be outside the brackets.