[[ Bug 20475 ]] Improve `Go to definition` for behaviors in use
This patch changes `seMatchingDefinitions` to use `revBehaviorUses`
to get the first user of a behavior and use that as the available
handler context. This requires some jiggery pokery with the descriptions
array so that the private handlers of the behavior script are included
and those of the user are not.

This patch also replaces the use of `revAvailableHandlers` with the
more efficient `revScriptDescription`.