[[Bug 19969]] Geometry Manager 'Left object' alignment issue
The PI for GM currently sets the ObjectSide for alignment to the same name as the side which aligns to the opposite side of the target.  If the target object scales, then the intended relation is not preserved.  The solution is to align the object to the nearest/opposite side.

If Btn A is on the left side of Btn B, then when you align the left side of Btn B to Btn A, it currently is calculated against the left side of Btn A.  When looking at it visually, it makes more sense to calculate against the right side of Btn A since that is the edge nearest to Btn B.

Fix adds a function to return the opposite side and changes the property assignment to use the new function.  This function will also be useful when enabling the proportional scaling of objects against the left side of the card (which the library code supports).