[Bug 17527] Unset proper geometry props when setting new ones (correction)
The original code for the new PI mistakenly cleared all "scale" props when "positioning" mode selections were made and vice versa.

The bug fix removed all of the code.  In actuality, it is proper to selectively remove props.  The GM library will not scale if move is selected.  With the current configuration of the PI someone could have move set and then set a proportional alignment, but it will not take effect.  This change expands on the original code with an additional optional parameter to the setValue handler with the key being set (used in 2 places).  This will ensure that the PI only displays settings that the GM library has a chance of using.

An example:  if you set "scaleLeft" or "scaleRight" (to card or object), then the "moveH" will be set to false to be sure that the new option takes effect.