[[Bug 17287]] Image filename property is not relativized with the PI
There is an option in preferences to "always use absolute file paths for images" but the PI does not change how it operates based on that setting.  This change modifies the file behavior to convert paths to relative if that setting is not made to use absolute.  It will impact both images and other media (anything in the PI that uses the file behavior).