[[Bug 17287]] Return absolute path if no common path elements
Add a check to return the absolute path if the original path is contained in the relative path.  This eliminates the series of `../` that would have been included otherwise.

On Windows this will prevent volume crossing since the first item is the drive letter.  On Mac and Linux it will only prevent volume crossing between the root file system and a mounted file system (but someone could create a relative link between 2 mounted file systems).  Checking for this on Mac would be fairly easy since the path starts with "/Volumes/" and item two is the drive name.  The path on an Ubuntu VM starts with "/media/", the second item is the user, and the third is the drive name.  I'm not sure if this is really something that would require a code solution.  Developers could easily do things that would be very difficult to detect like creating a relative link between local data and cloud data (Dropbox...).  It would be no worse than today where the absolute links would only work on the originating system.