[[Bug 13825/18097]] Profile Manager/Geometry Manager on Mobile
Profile Manager (PM) and Geometry Manager (GM) do not work on mobile due to missing library code in the standalone that is built.

GM depends on a handler present in the common library (REVTargetStack).  That library was not being loaded on mobile.  There is a good bit of code in that library that is not applicable to mobile.  This PR includes a new library of the common code that is pertinent to the mobile platform (revmobilelibrary).

PM depends on getprop/setprop handlers in the common library.  They are included in this new library.

The standalone settings excluded the profile manager option when iOS was selected.  That exclusion has been removed.

In this first commit, every handler name from the common library was left as a comment in the mobile library.  I used # for the ones that I'm confident could just be removed.  I've left the others with -- for discussion.  The intent would be to remove the comment lines before final merge.