[[Bug 20666]] Property Inspector / Geometry: Remove All button does not work
The first change adds a parameter to the setValue call (where the button click is handled) to remove all geometries and expands that handler to match (depends on an earlier PR that is merged already).

The second change is to add a command to clear geometry keys that are not needed.  If no geometry sides are set, then the other associated settings are removed.  This new command is called from the revIDESetGeometry handler.

The third change is to add checks to prevent the revCalculateGeometryDistances from calculating distances for sides that are not needed.