[[Bug 20684]] Dictionary navigation correction
Up/Down arrow keys navigate though the entire filtered list instead of the current search results.  Had to update previous/next functions to look for a current search term and use the cached search.

Moved the scrollTop lines to the displayEntryListGrep function (from dataSearch) since they need to be performed even when there is not a search term populated.  Also clear the cached search when the search term is blank.

Since JS is 0 based, changed the default index to 0 instead of 1.  This means that the actual first element gets displayed instead of the second when the dictionary is opened.  Changed null value for selected_entry_index to -1 to match data type.

Converted remaining spaces to tabs to make documents consistent throughout.