[[Bug 17287]] Relativize image filename property in the PI
There is an option in preferences to "always use absolute file paths for images" but the PI does not change how it operates based on that setting.  This change adds a setter for image filename that will relativize the path unless `cImageFileRelative` is false (default value is unset but shown as true in preferences, so check for it not being false vice checking for true).

Original PR for reference/prior discussion:
https://github.com/livecode/livecode-ide/pull/1817