[[ CodeResources ]] Fix revEnvironmentNonNativeBinariesPath
This patch fixes `revEnvironmentNonNativeBinariesPath` which had become
out of date since the gypification of build files. It now uses the bin
folder for each platform which can be downloaded from staging or copied
from another machine.

This patch also adds iOS, android and emscripten as platforms to the
function. In order to support iOS which has a childbuild folder per
SDK an additional pSDK parameter was added.