Fix initial setting of LCB editor pref
The `revIDEDeveloperExtensionEditScript` handler will open an LCB file in the default editor.  It uses `the result` to catch failures to open with a default editor and set a pref.  The problem is that after evaluating the next put (tEditor), the original value is lost.  This means that if the pref has not already been set, then the file won't get opened.

This change implements a new variable `tResult` to carry forward the result though the function.