[[ DataGrid 2 ]] Fix drags to cooperate with the mobile scroller.
Previously, a data grid's mobile scroller was getting in the way of
reorder and swipe actions.

To fix for reordering, we disable the scroller whenn the user clicks
on a reorder control (and re-enable it when the reorder completes).

Swipes are a little trickier. We track dragging on mouse down, but
disabling the scroller at this point will prevent all scrolling.
Instead we attempt to determine if the user is trying to drag the
row before disabling the scroller.